+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugfix
 * fix vertical bar cursor animation on empty command line
 * fix edge case while splitting the command line with formatting (better fix for [#379](https://github.com/jcubic/jquery.terminal/379))
+* fix option parser when using minus or double minus as argument
 
 ## 2.14.1
 ### Bugfix